requirements.txt
beautifulsoup4==4.10.0
progressbar33==2.4
pyfiglet==0.8.post1
requests==2.26.0
rich==11.0.0
